renamed old layout section file to position section
