+++ editor/src/components/inspector/inspector.tsx
@@ -77,7 +77,7 @@ import { FlexSection } from './flex-section'
 import { useDispatch } from '../editor/store/dispatch-context'
 import { styleStringInArray } from '../../utils/common-constants'
 import { SizingSection } from './sizing-section'
-import { PositionSection } from './sections/layout-section/layout-section'
+import { PositionSection } from './sections/layout-section/position-section'
 
 export interface ElementPathElement {
   name?: string
